django_core/requirements.txt
django==5.0.4
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
django-allauth==0.61.1
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
django-cors-headers==4.3.1
Pillow==10.4.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
prometheus-client==0.20.0